Moving function to tag package
